backend/requirements.txt
langchain 
langchain-community 
langchain-google-genai 
langchain-mcp-adapters
langgraph
mcp
fastmcp
fastapi
uvicorn
python-dotenv
tenacity
